updated package stress-ng
